feat: Score 계산 로직 분리 및 테스트 코드 작성
